# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "My DOC api",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/v1/destinations": {
            "get": {
                "tags": [
                    "Destinations"
                ],
                "summary": "Show all destinations",
                "operationId": "f7f86a72d87054a42a509a24b84a2e7e",
                "parameters": [
                    {
                        "name": "Accept-Language",
                        "in": "header",
                        "description": "Language preference for the response. Default is 'en'. Possible values: 'en', 'ru', 'kg', 'kz', 'tj', 'tm', 'uz'.",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "en",
                                "ru",
                                "kg",
                                "kz",
                                "tj",
                                "tm",
                                "uz"
                            ],
                            "example": "en"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Ok",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "integer",
                                            "example": 200
                                        },
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "slug": {
                                                        "type": "string",
                                                        "example": "1-kyrgyzstan"
                                                    },
                                                    "name": {
                                                        "type": "string",
                                                        "example": "Кыргызстан"
                                                    },
                                                    "image": {
                                                        "type": "string",
                                                        "example": "Test.png"
                                                    },
                                                    "destination": {
                                                        "type": "string",
                                                        "example": "Кыргызстан"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "integer",
                                            "example": 404
                                        },
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "error": {
                                            "type": "string",
                                            "example": "not_found."
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Not found."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/destinations/{destinations}": {
            "get": {
                "tags": [
                    "Destinations"
                ],
                "summary": "Show Destination",
                "operationId": "2ca5e656070cda24fa476d0f671e2dab",
                "parameters": [
                    {
                        "name": "destinations",
                        "in": "path",
                        "description": "ID Destinations",
                        "required": true,
                        "example": 1
                    },
                    {
                        "name": "Accept-Language",
                        "in": "header",
                        "description": "Language preference for the response. Default is 'en'. Possible values: 'en', 'ru', 'kg', 'kz', 'tj', 'tm', 'uz'.",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "en",
                                "ru",
                                "kg",
                                "kz",
                                "tj",
                                "tm",
                                "uz"
                            ],
                            "example": "en"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Ok",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "integer",
                                            "example": 200
                                        },
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "properties": {
                                                "slug": {
                                                    "type": "string",
                                                    "example": "1-kyrgyzstan"
                                                },
                                                "name": {
                                                    "type": "string",
                                                    "example": "Кыргызстан"
                                                },
                                                "image": {
                                                    "type": "string",
                                                    "example": "Test.png"
                                                },
                                                "destination": {
                                                    "type": "string",
                                                    "example": "Кыргызстан"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "integer",
                                            "example": 404
                                        },
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "error": {
                                            "type": "string",
                                            "example": "not_found."
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Not found."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/destinations/travel/{destinations}": {
            "get": {
                "tags": [
                    "Destinations"
                ],
                "summary": "Show Travel-Destinations in Destination",
                "operationId": "089b8060d26d5e659eb33ea49e9d35e7",
                "parameters": [
                    {
                        "name": "destinations",
                        "in": "path",
                        "description": "ID destinations",
                        "required": true,
                        "example": 1
                    },
                    {
                        "name": "Accept-Language",
                        "in": "header",
                        "description": "Language preference for the response. Default is 'en'. Possible values: 'en', 'ru', 'kg', 'kz', 'tj', 'tm', 'uz'.",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "en",
                                "ru",
                                "kg",
                                "kz",
                                "tj",
                                "tm",
                                "uz"
                            ],
                            "example": "en"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Ok",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "integer",
                                            "example": 200
                                        },
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "slug": {
                                                        "type": "string",
                                                        "example": "1-altyn-arashan"
                                                    },
                                                    "name": {
                                                        "type": "string",
                                                        "example": "Алатын-Арашан"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "integer",
                                            "example": 404
                                        },
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "error": {
                                            "type": "string",
                                            "example": "not_found."
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Not found."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
...
}